fix(fsmv2): address code review issues for production-quality code
Code review findings and fixes:

1. Deep clone for VariableBundle (variables.go)
   - Implement JSON marshal/unmarshal based deep cloning
   - Nested maps/slices are now fully independent after Clone()
   - Add tests verifying nested map modifications don't affect original

2. Add ChildSpec.Clone() method (childspec.go)
   - Deep copy UserSpec using its Clone() method
   - Deep copy ChildStartStates slice
   - Add tests for nil/empty slice handling

3. Document example worker test patterns (state_stopped.go, state_running.go, doc.go)
   - Add prominent warnings that these are EXAMPLE/TEST workers only
   - Document why package-level vars are acceptable here
   - List recommended alternatives for production workers
   - Explain anti-pattern context for future developers

4. Fix GetPulledMessages documentation (dependencies.go)
   - Clarify shallow copy semantics (slice copied, pointers shared)
   - Document that callers should treat messages as read-only

5. Fix test for JSON number handling (variable_injection_test.go)
   - JSON marshal/unmarshal converts int to float64
   - Update test expectation to use float64(502)

+++ umh-core/pkg/fsmv2/supervisor/variable_injection_test.go
@@ -451,9 +451,10 @@ var _ = Describe("Variable Injection", func() {
 	Describe("User Variables Preservation", func() {
 		It("should preserve existing User variables during injection", func() {
 			// Set up userSpec with existing User variables
+			// Note: PORT uses float64 because JSON deep cloning converts all numbers to float64
 			existingUserVars := map[string]any{
 				"IP":   "192.168.1.100",
-				"PORT": 502,
+				"PORT": float64(502),
 			}
 
 			userSpec := config.UserSpec{
